Save sync routing setting in port state
This eliminates the need to look up the setting in the application
environment for each outgoing packet.  The setting is read when each
port is initialised.

There is one remaining case where the application environment is
consulted: when a packet is sent by a packet output action that
specifies the port as 'table'.  In that case, we don't have access to
the state of a port process.